moving Rules integration to a submodule
