+++ modules/civicrm_entity_rules/src/Plugin/TypedDataFilter/FirstLastFilter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\civicrm_entity\Plugin\TypedDataFilter;
+namespace Drupal\civicrm_entity_rules\Plugin\TypedDataFilter;
 
 use Drupal\Core\Render\BubbleableMetadata;
 use Drupal\Core\TypedData\DataDefinition;
@@ -11,11 +11,11 @@ use Drupal\typed_data\DataFilterBase;
  * A data filter providing a default value if no value is set.
  *
  * @DataFilter(
- *   id = "initialdotlast",
- *   label = @Translation("Drupal Username : initial.lastname."),
+ *   id = "FirstLast",
+ *   label = @Translation("Drupal Username : FirstnameLastname."),
  * )
  */
-class InitialDotLastFilter extends DataFilterBase {
+class FirstLastFilter extends DataFilterBase {
 
   /**
    * {@inheritdoc}
@@ -40,13 +40,7 @@ class InitialDotLastFilter extends DataFilterBase {
    * {@inheritdoc}
    */
   public function filter(DataDefinitionInterface $definition, $value, array $arguments, BubbleableMetadata $bubbleable_metadata = NULL) {
-    $c = preg_match_all("/(?<=\b)[a-z]/i", ($value->get('first_name')->getString()), $m);
-    if ($c > 0) {
-      $login = strtolower(implode('', $m[0])) . '.' . strtolower($value->get('last_name')->getString());
-    }
-    else {
-      $login = strtolower($value->get('last_name')->getString());
-    }
+    $login = str_replace(' ', '', ucfirst(strtolower($value->get('first_name')->getString()))) . ucfirst(strtolower($value->get('last_name')->getString()));
 
     return filter_var($login, FILTER_SANITIZE_EMAIL);
   }
